migrate estimate_inplane_rotations, complete_third_row_to_rot to commonline utils.
